+++ sf_examples/mujoco_examples/enjoy_mujoco.py
@@ -1,14 +1,13 @@
 import sys
 
-from sample_factory.cfg.arguments import parse_args
 from sample_factory.enjoy import enjoy
-from sf_examples.mujoco_examples.train_mujoco import register_mujoco_components
+from sf_examples.mujoco_examples.train_mujoco import register_mujoco_components, parse_mujoco_cfg
 
 
 def main():
     """Script entry point."""
     register_mujoco_components()
-    cfg = parse_args(evaluation=True)
+    cfg = parse_mujoco_cfg(evaluation=True)
     status = enjoy(cfg)
     return status
