+++ sf_examples/mujoco_examples/train_mujoco.py
@@ -1,26 +1,29 @@
 import sys
 
-from sample_factory.algo.utils.context import global_env_registry
-from sample_factory.cfg.arguments import parse_args
+from sample_factory.cfg.arguments import parse_full_cfg, parse_sf_args
+from sample_factory.envs.env_utils import register_env
 from sample_factory.train import run_rl
+from sf_examples.mujoco_examples.mujoco.mujoco_params import add_mujoco_env_args, mujoco_override_defaults
+from sf_examples.mujoco_examples.mujoco.mujoco_utils import MUJOCO_ENVS, make_mujoco_env
 
 
 def register_mujoco_components():
-    from sf_examples.mujoco_examples.mujoco_params import add_mujoco_env_args, mujoco_override_defaults
-    from sf_examples.mujoco_examples.mujoco_utils import make_mujoco_env
+    for env in MUJOCO_ENVS:
+        register_env(env.name, make_mujoco_env)
 
-    global_env_registry().register_env(
-        env_name_prefix="mujoco_",
-        make_env_func=make_mujoco_env,
-        add_extra_params_func=add_mujoco_env_args,
-        override_default_params_func=mujoco_override_defaults,
-    )
+
+def parse_mujoco_cfg(argv=None, evaluation=False):
+    parser, partial_cfg = parse_sf_args(argv=argv, evaluation=evaluation)
+    add_mujoco_env_args(partial_cfg.env, parser)
+    mujoco_override_defaults(partial_cfg.env, parser)
+    final_cfg = parse_full_cfg(parser, argv)
+    return final_cfg
 
 
 def main():
     """Script entry point."""
     register_mujoco_components()
-    cfg = parse_args()
+    cfg = parse_mujoco_cfg()
     status = run_rl(cfg)
     return status
