+++ tests/examples/test_mujoco.py
@@ -4,18 +4,16 @@ from os.path import isdir
 import pytest
 
 from sample_factory.algo.utils.misc import ExperimentStatus
-from sample_factory.cfg.arguments import parse_args
+from sf_examples.mujoco_examples.train_mujoco import register_mujoco_components, parse_mujoco_cfg
 from sample_factory.train import run_rl
 from sample_factory.utils.utils import experiment_dir, log
-from sf_examples.mujoco_examples.mujoco_utils import mujoco_available
+from sf_examples.mujoco_examples.mujoco.mujoco_utils import mujoco_available
 
 
 @pytest.mark.skipif(not mujoco_available(), reason="mujoco not installed")
 class TestMujoco:
     @pytest.fixture(scope="class", autouse=True)
     def register_mujoco_fixture(self):
-        from sf_examples.mujoco_examples.train_mujoco import register_mujoco_components
-
         register_mujoco_components()
 
     @staticmethod
@@ -34,7 +32,7 @@ class TestMujoco:
 
         experiment_name = "test_" + env
 
-        cfg = parse_args(argv=["--algo=APPO", f"--env={env}", f"--experiment={experiment_name}"])
+        cfg = parse_mujoco_cfg(argv=["--algo=APPO", f"--env={env}", f"--experiment={experiment_name}"])
         cfg.serial_mode = serial_mode
         cfg.async_rl = async_rl
         cfg.batched_sampling = batched_sampling
